prepared for heroku deployment

+++ middleware/cache.js
@@ -1,18 +1,22 @@
 const {redis} = require('../src/util/redis');
 
 const checkCache = async (req, res, next, key) => {
-    const cache = await redis.get(key);
-    if (cache) return res.send(JSON.parse(cache));
+    if (redis) {
+        const cache = await redis.get(key);
+        if (cache) return res.send(JSON.parse(cache));
+    }
     next();
 };
 
-const setCache = (key, data,expiry) => {
-    redis.set(key, JSON.stringify(data));
-    if (expiry) redis.expire(key,expiry);
+const setCache = (key, data, expiry) => {
+    if(redis) {
+        redis.set(key, JSON.stringify(data));
+        if (expiry) redis.expire(key, expiry);
+    }
 };
 
 const flushCache = () => {
-    redis.flushall();
+    if(redis) redis.flushall();
 };
 
 module.exports = {checkCache, setCache, flushCache};
